Rename .h -> .hpp
