+++ src/token.cpp
@@ -1,4 +1,4 @@
-#include "token.h"
+#include "token.hpp"
 
 #include <iomanip>
 #include <iostream>
